Add foundation source file.
Functionality in this file needs to be available globally as everything is likely to depend on it.
Even if some functionality would make sense to exist in another source. If it has common use and is lightweight then it will be added here to avoid having to include extra functionality unnecessarily.